tests/wpt/tests/remote-playback/disable-remote-playback-cancel-watch-availability-throws.html

<!DOCTYPE html>
<html>
  <title>
    Test that calling cancelWatchAvailability() when disableRemotePlayback attribute is set throws InvalidStateError
  </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/media.js"></script>
  <script>
    async_test(t => {
      let v = document.createElement("video");
      v.src = getVideoURI("/media/movie_5");

      v.remote
        .watchAvailability(() => {})
        .then(id => {
          v.disableRemotePlayback = true;
          v.remote.cancelWatchAvailability(id).then(
            t.unreached_func(),
            t.step_func(e => {
              assert_equals(e.name, "InvalidStateError");
              v.remote.cancelWatchAvailability().then(
                t.unreached_func(),
                t.step_func_done(e => {
                  assert_equals(e.name, "InvalidStateError");
                })
              );
            })
          );
        }, t.unreached_func());
    }, "Test that calling cancelWatchAvailability() when disableRemotePlayback attribute is set throws InvalidStateError.");
  </script>
</html>
